requirements.txt
openai-whisper
spacy
pyttsx3
python-dotenv
transformers
torch
nltk
PyQt5
azure-ai-inference
azure-core
azure-identity
jedi
watchdog
pytest
coverage
pyyaml
radon
pylint
docstring-parser
mdutils
memory-profiler
psutil
stack_data
toml
numpy
packaging
webrtcvad
scipy
rich
azure-cognitiveservices-vision-computervision
duckduckgo_search
wolframalpha
Pillow
pyaudio
python-dotenv
whisper
datetime